feature: add MobileNumberValidator interface
